Move the call back server mocks to sane locations & names
& delete the no longer relevant KBaseJobService spec
