+++ src/java/us/kbase/test/sdk/initializer/CallbackServerMock.java
@@ -1,8 +1,9 @@
-package us.kbase.kbasejobservice;
+package us.kbase.test.sdk.initializer;
 
 import java.io.File;
 import java.io.IOException;
 import java.io.OutputStream;
+import java.io.PrintWriter;
 
 import us.kbase.auth.AuthToken;
 import us.kbase.common.service.JacksonTupleModule;
@@ -11,7 +12,6 @@ import us.kbase.common.service.JsonServerSyslog;
 
 //BEGIN_HEADER
 
-// There's a similar class in us.kbase.test.sdk.initializer.ExecEngineMock
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -23,13 +23,16 @@ import java.util.regex.Pattern;
 
 import javax.servlet.http.HttpServletResponse;
 
+import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang.exception.ExceptionUtils;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.ImmutableMap;
 
 import us.kbase.common.service.UObject;
+import us.kbase.mobu.util.DirUtils;
 import us.kbase.mobu.util.ProcessHelper;
+import us.kbase.test.sdk.scripts.TestConfigHelper;
 
 //END_HEADER
 
@@ -38,35 +41,30 @@ import us.kbase.mobu.util.ProcessHelper;
  * <pre>
  * </pre>
  */
-public class KBaseJobServiceServer extends JsonServerServlet {
+public class CallbackServerMock extends JsonServerServlet {
     private static final long serialVersionUID = 1L;
 
     //BEGIN_CLASS_HEADER
     private int lastJobId = 0;
-    private final Map<String, Map<String, Object>> results = new LinkedHashMap<>();
-    private File binDir = null;
-    private File tempDir = null;
+    private String kbaseEndpoint = null;
+    private Map<String, String> jobToModule = new LinkedHashMap<String, String>();
+    private Map<String, File> jobToWorkDir = new LinkedHashMap<String, File>();
+    private Map<String, Thread> jobToWorker = new LinkedHashMap<String, Thread>();
+    private Map<String, Map<String, Object>> jobToResults = new LinkedHashMap<>();
+    private Map<String, String> moduleToDockerImage = new LinkedHashMap<String, String>();
+    private Map<String, File> moduleToRepoDir = new LinkedHashMap<String, File>();
+    private static boolean debugCheckJobTimes = false;
+    private Long lastCheckJobAccessTime = null;
     
-    public KBaseJobServiceServer withBinDir(File binDir) {
-        this.binDir = binDir;
+    public CallbackServerMock withModule(String moduleName, String dockerImage, File repoDir) {
+        moduleToDockerImage.put(moduleName, dockerImage);
+        moduleToRepoDir.put(moduleName, repoDir);
         return this;
     }
     
-    public KBaseJobServiceServer withTempDir(File tempDir) {
-        this.tempDir = tempDir;
+    public CallbackServerMock withKBaseEndpoint(String endpoint) {
+        this.kbaseEndpoint = endpoint;
         return this;
-    }
-    
-    private String getBinScript(String scriptName) {
-        File ret = null;
-        if (binDir == null) {
-            ret = new File(".", scriptName);
-            if (!ret.exists())
-                return scriptName;
-        } else {
-            ret = new File(binDir, scriptName);
-        }
-        return ret.getAbsolutePath();
     }
     
     protected void processRpcCall(RpcCallData rpcCallData, String token, JsonServerSyslog.RpcInfo info, 
@@ -98,7 +96,11 @@ public class KBaseJobServiceServer extends JsonServerServlet {
                     result.add(runJob(runJobParams, t));
                 } else if (rpcName.endsWith("._check_job") && paramsList.size() == 1) {
                     String jobId = paramsList.get(0).asClassInstance(String.class);
-                    Map<String, Object> fjp = results.get(jobId);
+                    if (lastCheckJobAccessTime != null && debugCheckJobTimes) {
+                        System.out.println("ExecEngineMock.checkJob: time = " + (System.currentTimeMillis() - lastCheckJobAccessTime));
+                    }
+                    lastCheckJobAccessTime = System.currentTimeMillis();
+                    Map<String, Object> fjp = jobToResults.get(jobId);
                     if (fjp != null && fjp.get("error") != null) {
                         Map<String, Object> ret = new LinkedHashMap<String, Object>();
                         ret.put("version", "1.1");
@@ -181,10 +183,24 @@ public class KBaseJobServiceServer extends JsonServerServlet {
             inner.write(b, off, len);
         }
     }
+    
+    public void waitAndCleanAllJobs() throws Exception {
+        for (String jobId : new ArrayList<String>(jobToModule.keySet())) {
+            String moduleName = jobToModule.get(jobId);
+            Thread t = jobToWorker.get(jobId);
+            t.join();
+            jobToModule.remove(jobId);
+            jobToResults.remove(jobId);
+            jobToWorkDir.remove(jobId);
+            jobToWorker.remove(jobId);
+            moduleToDockerImage.remove(moduleName);
+            moduleToRepoDir.remove(moduleName);
+        }
+    }
     //END_CLASS_HEADER
 
-    public KBaseJobServiceServer() throws Exception {
-        super("KBaseJobService");
+    public CallbackServerMock() throws Exception {
+        super("NarrativeJobService");
         //BEGIN_CONSTRUCTOR
         //END_CONSTRUCTOR
     }
@@ -193,41 +209,53 @@ public class KBaseJobServiceServer extends JsonServerServlet {
         lastJobId++;
         final String jobId = "" + lastJobId;
         final AuthToken token = authPart;
-        final File jobDir = new File(tempDir, "job_" + jobId);
+        String moduleName = ((String)params.get("method")).split(Pattern.quote("."))[0];
+        jobToModule.put(jobId, moduleName);
+        File moduleDir = moduleToRepoDir.get(moduleName);
+        File testLocalDir = new File(moduleDir, "test_local");
+        final File jobDir = new File(testLocalDir, "job_" + jobId);
         if (!jobDir.exists())
             jobDir.mkdirs();
+        jobToWorkDir.put(jobId, jobDir);
         File jobFile = new File(jobDir, "job.json");
         UObject.getMapper().writeValue(jobFile, params);
-        new Thread(new Runnable() {
+        final String dockerImage = moduleToDockerImage.get(moduleName);
+        Map<String, Object> rpc = new LinkedHashMap<String, Object>();
+        rpc.put("version", "1.1");
+        rpc.put("method", params.get("method"));
+        rpc.put("params", params.get("params"));
+        File tokenFile = new File(jobDir, "token");
+        FileUtils.writeStringToFile(tokenFile, token.getToken());
+        File inputFile = new File(jobDir, "input.json");
+        UObject.getMapper().writeValue(inputFile, rpc);
+        final File outputFile = new File(jobDir, "output.json");
+        final String containerName = "test_" + moduleName.toLowerCase() + "_" + 
+                System.currentTimeMillis();
+        File runDockerSh = new File(testLocalDir, "run_docker.sh");
+        final String runDockerPath = DirUtils.getFilePath(runDockerSh);
+        Thread t = new Thread(new Runnable() {
             @Override
             public void run() {
                 try {
-                    final File jobDir = new File(tempDir, "job_" + jobId);
-                    final File jobFile = new File(jobDir, "job.json");
-                    @SuppressWarnings("unchecked")
-                    final Map<String, Object> job = UObject.getMapper().readValue(
-                            jobFile, Map.class
-                    );
-                    String serviceName = ((String)job.get("method")).split("\\.")[0];
-                    if (!jobDir.exists()) {
-                        jobDir.mkdirs();
+                    ProcessHelper.cmd("bash", runDockerPath, "run", "-v", 
+                            jobDir.getCanonicalPath() + ":/kb/module/work", 
+                            "--name", containerName, "-e", "KBASE_ENDPOINT=" + kbaseEndpoint,
+                            "-e", "AUTH_SERVICE_URL=" + TestConfigHelper.getAuthServiceUrl(),
+                            "-e", "AUTH_SERVICE_URL_ALLOW_INSECURE=" + 
+                                    TestConfigHelper.getAuthServiceUrlInsecure(),
+                                    dockerImage, "async").exec(jobDir, 
+                                            null, (PrintWriter)null, null);
+                    if (!outputFile.exists()) {
+                        ProcessHelper.cmd("bash", runDockerPath, "logs", containerName).exec(jobDir);
+                        throw new IllegalStateException("Output file wasn't created");
                     }
-                    Map<String, Object> rpc = new LinkedHashMap<String, Object>();
-                    rpc.put("version", "1.1");
-                    rpc.put("method", job.get("method"));
-                    rpc.put("params", job.get("params"));
-                    File inputFile = new File(jobDir, "input.json");
-                    UObject.getMapper().writeValue(inputFile, rpc);
-                    String scriptFilePath = getBinScript("run_" + serviceName + "_async_job.sh");
-                    File outputFile = new File(jobDir, "output.json");
-                    ProcessHelper.cmd("bash", scriptFilePath, inputFile.getCanonicalPath(),
-                            outputFile.getCanonicalPath(), token.getToken()).exec(jobDir);
                     @SuppressWarnings("unchecked")
                     Map<String, Object> result = UObject.getMapper().readValue(
                             outputFile, Map.class
                     );
-                    results.put(jobId, result);
+                    jobToResults.put(jobId, result);
                 } catch (Exception ex) {
+                    ex.printStackTrace();
                     final Map<String, Object> result = ImmutableMap.of("error",
                             (Object) ImmutableMap.of(
                                     "code", -1L,
@@ -235,10 +263,19 @@ public class KBaseJobServiceServer extends JsonServerServlet {
                                     "message", "Job service side error: " + ex.getMessage()
                             )
                     );
-                    results.put(jobId, result);
+                    jobToResults.put(jobId, result);
+                } finally {
+                    try {
+                        ProcessHelper.cmd("bash", runDockerPath, "rm", "-v", "-f", 
+                                containerName).exec(jobDir, null, (PrintWriter)null, null);
+                    } catch (Exception ex) {
+                        System.out.println("Error deleting container [" + containerName + "]: " + ex.getMessage());
+                    }
                 }
             }
-        }).start();
+        });
+        t.start();
+        jobToWorker.put(jobId, t);
         return jobId;
     }
 }
