+++ src/java/us/kbase/test/sdk/scripts/CallbackServerMock.java
@@ -1,9 +1,8 @@
-package us.kbase.test.sdk.initializer;
+package us.kbase.test.sdk.scripts;
 
 import java.io.File;
 import java.io.IOException;
 import java.io.OutputStream;
-import java.io.PrintWriter;
 
 import us.kbase.auth.AuthToken;
 import us.kbase.common.service.JacksonTupleModule;
@@ -12,6 +11,7 @@ import us.kbase.common.service.JsonServerSyslog;
 
 //BEGIN_HEADER
 
+// There's a similar class in us.kbase.test.sdk.initializer.ExecEngineMock
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -23,16 +23,13 @@ import java.util.regex.Pattern;
 
 import javax.servlet.http.HttpServletResponse;
 
-import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang.exception.ExceptionUtils;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.ImmutableMap;
 
 import us.kbase.common.service.UObject;
-import us.kbase.mobu.util.DirUtils;
 import us.kbase.mobu.util.ProcessHelper;
-import us.kbase.test.sdk.scripts.TestConfigHelper;
 
 //END_HEADER
 
@@ -41,30 +38,35 @@ import us.kbase.test.sdk.scripts.TestConfigHelper;
  * <pre>
  * </pre>
  */
-public class ExecEngineMock extends JsonServerServlet {
+public class CallbackServerMock extends JsonServerServlet {
     private static final long serialVersionUID = 1L;
 
     //BEGIN_CLASS_HEADER
     private int lastJobId = 0;
-    private String kbaseEndpoint = null;
-    private Map<String, String> jobToModule = new LinkedHashMap<String, String>();
-    private Map<String, File> jobToWorkDir = new LinkedHashMap<String, File>();
-    private Map<String, Thread> jobToWorker = new LinkedHashMap<String, Thread>();
-    private Map<String, Map<String, Object>> jobToResults = new LinkedHashMap<>();
-    private Map<String, String> moduleToDockerImage = new LinkedHashMap<String, String>();
-    private Map<String, File> moduleToRepoDir = new LinkedHashMap<String, File>();
-    private static boolean debugCheckJobTimes = false;
-    private Long lastCheckJobAccessTime = null;
+    private final Map<String, Map<String, Object>> results = new LinkedHashMap<>();
+    private File binDir = null;
+    private File tempDir = null;
     
-    public ExecEngineMock withModule(String moduleName, String dockerImage, File repoDir) {
-        moduleToDockerImage.put(moduleName, dockerImage);
-        moduleToRepoDir.put(moduleName, repoDir);
+    public CallbackServerMock withBinDir(File binDir) {
+        this.binDir = binDir;
         return this;
     }
     
-    public ExecEngineMock withKBaseEndpoint(String endpoint) {
-        this.kbaseEndpoint = endpoint;
+    public CallbackServerMock withTempDir(File tempDir) {
+        this.tempDir = tempDir;
         return this;
+    }
+    
+    private String getBinScript(String scriptName) {
+        File ret = null;
+        if (binDir == null) {
+            ret = new File(".", scriptName);
+            if (!ret.exists())
+                return scriptName;
+        } else {
+            ret = new File(binDir, scriptName);
+        }
+        return ret.getAbsolutePath();
     }
     
     protected void processRpcCall(RpcCallData rpcCallData, String token, JsonServerSyslog.RpcInfo info, 
@@ -96,11 +98,7 @@ public class ExecEngineMock extends JsonServerServlet {
                     result.add(runJob(runJobParams, t));
                 } else if (rpcName.endsWith("._check_job") && paramsList.size() == 1) {
                     String jobId = paramsList.get(0).asClassInstance(String.class);
-                    if (lastCheckJobAccessTime != null && debugCheckJobTimes) {
-                        System.out.println("ExecEngineMock.checkJob: time = " + (System.currentTimeMillis() - lastCheckJobAccessTime));
-                    }
-                    lastCheckJobAccessTime = System.currentTimeMillis();
-                    Map<String, Object> fjp = jobToResults.get(jobId);
+                    Map<String, Object> fjp = results.get(jobId);
                     if (fjp != null && fjp.get("error") != null) {
                         Map<String, Object> ret = new LinkedHashMap<String, Object>();
                         ret.put("version", "1.1");
@@ -183,24 +181,10 @@ public class ExecEngineMock extends JsonServerServlet {
             inner.write(b, off, len);
         }
     }
-    
-    public void waitAndCleanAllJobs() throws Exception {
-        for (String jobId : new ArrayList<String>(jobToModule.keySet())) {
-            String moduleName = jobToModule.get(jobId);
-            Thread t = jobToWorker.get(jobId);
-            t.join();
-            jobToModule.remove(jobId);
-            jobToResults.remove(jobId);
-            jobToWorkDir.remove(jobId);
-            jobToWorker.remove(jobId);
-            moduleToDockerImage.remove(moduleName);
-            moduleToRepoDir.remove(moduleName);
-        }
-    }
     //END_CLASS_HEADER
 
-    public ExecEngineMock() throws Exception {
-        super("NarrativeJobService");
+    public CallbackServerMock() throws Exception {
+        super("KBaseJobService");
         //BEGIN_CONSTRUCTOR
         //END_CONSTRUCTOR
     }
@@ -209,53 +193,41 @@ public class ExecEngineMock extends JsonServerServlet {
         lastJobId++;
         final String jobId = "" + lastJobId;
         final AuthToken token = authPart;
-        String moduleName = ((String)params.get("method")).split(Pattern.quote("."))[0];
-        jobToModule.put(jobId, moduleName);
-        File moduleDir = moduleToRepoDir.get(moduleName);
-        File testLocalDir = new File(moduleDir, "test_local");
-        final File jobDir = new File(testLocalDir, "job_" + jobId);
+        final File jobDir = new File(tempDir, "job_" + jobId);
         if (!jobDir.exists())
             jobDir.mkdirs();
-        jobToWorkDir.put(jobId, jobDir);
         File jobFile = new File(jobDir, "job.json");
         UObject.getMapper().writeValue(jobFile, params);
-        final String dockerImage = moduleToDockerImage.get(moduleName);
-        Map<String, Object> rpc = new LinkedHashMap<String, Object>();
-        rpc.put("version", "1.1");
-        rpc.put("method", params.get("method"));
-        rpc.put("params", params.get("params"));
-        File tokenFile = new File(jobDir, "token");
-        FileUtils.writeStringToFile(tokenFile, token.getToken());
-        File inputFile = new File(jobDir, "input.json");
-        UObject.getMapper().writeValue(inputFile, rpc);
-        final File outputFile = new File(jobDir, "output.json");
-        final String containerName = "test_" + moduleName.toLowerCase() + "_" + 
-                System.currentTimeMillis();
-        File runDockerSh = new File(testLocalDir, "run_docker.sh");
-        final String runDockerPath = DirUtils.getFilePath(runDockerSh);
-        Thread t = new Thread(new Runnable() {
+        new Thread(new Runnable() {
             @Override
             public void run() {
                 try {
-                    ProcessHelper.cmd("bash", runDockerPath, "run", "-v", 
-                            jobDir.getCanonicalPath() + ":/kb/module/work", 
-                            "--name", containerName, "-e", "KBASE_ENDPOINT=" + kbaseEndpoint,
-                            "-e", "AUTH_SERVICE_URL=" + TestConfigHelper.getAuthServiceUrl(),
-                            "-e", "AUTH_SERVICE_URL_ALLOW_INSECURE=" + 
-                                    TestConfigHelper.getAuthServiceUrlInsecure(),
-                                    dockerImage, "async").exec(jobDir, 
-                                            null, (PrintWriter)null, null);
-                    if (!outputFile.exists()) {
-                        ProcessHelper.cmd("bash", runDockerPath, "logs", containerName).exec(jobDir);
-                        throw new IllegalStateException("Output file wasn't created");
+                    final File jobDir = new File(tempDir, "job_" + jobId);
+                    final File jobFile = new File(jobDir, "job.json");
+                    @SuppressWarnings("unchecked")
+                    final Map<String, Object> job = UObject.getMapper().readValue(
+                            jobFile, Map.class
+                    );
+                    String serviceName = ((String)job.get("method")).split("\\.")[0];
+                    if (!jobDir.exists()) {
+                        jobDir.mkdirs();
                     }
+                    Map<String, Object> rpc = new LinkedHashMap<String, Object>();
+                    rpc.put("version", "1.1");
+                    rpc.put("method", job.get("method"));
+                    rpc.put("params", job.get("params"));
+                    File inputFile = new File(jobDir, "input.json");
+                    UObject.getMapper().writeValue(inputFile, rpc);
+                    String scriptFilePath = getBinScript("run_" + serviceName + "_async_job.sh");
+                    File outputFile = new File(jobDir, "output.json");
+                    ProcessHelper.cmd("bash", scriptFilePath, inputFile.getCanonicalPath(),
+                            outputFile.getCanonicalPath(), token.getToken()).exec(jobDir);
                     @SuppressWarnings("unchecked")
                     Map<String, Object> result = UObject.getMapper().readValue(
                             outputFile, Map.class
                     );
-                    jobToResults.put(jobId, result);
+                    results.put(jobId, result);
                 } catch (Exception ex) {
-                    ex.printStackTrace();
                     final Map<String, Object> result = ImmutableMap.of("error",
                             (Object) ImmutableMap.of(
                                     "code", -1L,
@@ -263,19 +235,10 @@ public class ExecEngineMock extends JsonServerServlet {
                                     "message", "Job service side error: " + ex.getMessage()
                             )
                     );
-                    jobToResults.put(jobId, result);
-                } finally {
-                    try {
-                        ProcessHelper.cmd("bash", runDockerPath, "rm", "-v", "-f", 
-                                containerName).exec(jobDir, null, (PrintWriter)null, null);
-                    } catch (Exception ex) {
-                        System.out.println("Error deleting container [" + containerName + "]: " + ex.getMessage());
-                    }
+                    results.put(jobId, result);
                 }
             }
-        });
-        t.start();
-        jobToWorker.put(jobId, t);
+        }).start();
         return jobId;
     }
 }
